SeedBuilder: Document and improve SeedBuilder syntax
